+++ selfinstaller/src/steps.rs
@@ -3,3 +3,5 @@ pub use directory::directory;
 pub mod file;
 pub use file::{executable, file};
 pub mod systemd;
+#[cfg(windows)]
+pub mod windows;